Fix LIRC::Commands.serialize_command_name under ruby 2.5
I went back-and-forth on this for a bit - I think the best solution is
to only run mutant on the minimum ruby version I target - that way I can
ensure it doesn't perform mutations that require newer syntax.

This is definitely the simplest, fastest, least semantically-rich way to
chop the first character off a string in ruby 2.5. ruby 2.6 seems
impossible to get 100% mutation coverage on - it'll suggest [1..nil] if
you have [1...nil], and vice-versa.

Lucky for us, that's illegal in ruby 2.5